Cleaning house with refactoring and corrections.
Main.h now the project header, with main.c includes stored in the source file. Added battcheck and some state machine logic. Corrected some various functions and added missing tags.

+++ Discharger_0/Src/ADC.c
@@ -190,7 +190,7 @@ void ADC1_Cycle_Stop(void){
 /*	Start an ADC1 conversion							*/
 /*	Args: Channel to run, addr of destination buffer	*/
 
-uint8_t ADC1_Start_Conv(uint8_t ch, uint32_t *buffer){
+uint8_t ADC1_Start_Conv(uint8_t ch, uint16_t *buffer){
 
 	if (!((ADC1 -> CR >> 2) & 0x1)) {		// Check for no Conver in progress
